Move hardware initialisation to package module
