+++ homeassistant/components/hardware/models.py
@@ -3,20 +3,29 @@
 from __future__ import annotations
 
 from dataclasses import dataclass
-from typing import TYPE_CHECKING, Protocol
+from typing import Protocol
 
-from homeassistant.core import HomeAssistant, callback
+import psutil_home_assistant as ha_psutil
 
-if TYPE_CHECKING:
-    from .websocket_api import SystemStatus
+from homeassistant.components import websocket_api
+from homeassistant.core import CALLBACK_TYPE, HomeAssistant, callback
 
 
 @dataclass
 class HardwareData:
     """Hardware data."""
 
-    hardware_platform: dict[str, HardwareProtocol] = None  # type: ignore[assignment]
-    system_status: SystemStatus = None  # type: ignore[assignment]
+    hardware_platform: dict[str, HardwareProtocol]
+    system_status: SystemStatus
+
+
+@dataclass(slots=True)
+class SystemStatus:
+    """System status."""
+
+    ha_psutil: ha_psutil
+    remove_periodic_timer: CALLBACK_TYPE | None
+    subscribers: set[tuple[websocket_api.ActiveConnection, int]]
 
 
 @dataclass(slots=True)
